Fix relative links (#11)
## 💸 TL;DR
Due to [how Docusaurus resolves
links](https://docusaurus.io/docs/markdown-features/links), if we're
linking to an internal Markdown file, we should be sure to include the
extension. This ensures that when looking at a versioned doc, or looking
at a page with a trailing slash, that the URLs still resolve correctly,
as Docusaurus will replace the relative path with a correct, absolute
path at build time.

## 🧪 Testing Steps / Validation
`yarn build && yarn serve` locally worked great for me 👍